Refactor: Create package util and file for linear trajectory
Create a package util that will contain functions not related to other packages. In addition, move class LinearTrajectory to its own file.
