+++ src/plotter/simulation_plotter.py
@@ -9,7 +9,7 @@ from os.path import exists, dirname
 from constants import DELTA_T, CONTROLLER, PATH_TO_EXPORT_DATA
 from constants import STEPS
 from plotter import Plotter, PlotData, get_error
-from trajectory.trajectory import create_trajectory
+from util.util import create_trajectory
 
 
 class SimulationPlotter(Plotter):
